feat: Correções

+++ flor-de-lotus/flor-de-lotus/src/main/java/flor_de_lotus/endereco/Endereco.java
@@ -15,7 +15,7 @@ import lombok.Setter;
 public class Endereco {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long idEndereco;
+    private Integer idEndereco;
     private String logradouro;
     private String numero;
     private String complemento;
